refactor(backend): restructure serial communication system
- Replace select with msvcrt for Windows compatibility
- Add proper process termination and cleanup
- Move serial handler to shared location
- Improve data parsing and error handling
- Add position data streaming support
- Update IPC event handling

BREAKING CHANGE: Serial handler has been moved to shared location and requires Windows-specific msvcrt module
